requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
psycopg2>=2.9
PyPDF2>=3.0.0
langchain
langchain-groq
PyMuPDF
python-dotenv
django-grappelli
celery
redis
django-redis
